split the sign module into multiple files
